Suppress image warning.
