--- docs/_ext/__init__.py
@@ -1,1 +1,0 @@
-from environmentSetup import supress_nonlocal_image_warn
