+++ docs/conf.py
@@ -258,5 +258,14 @@ texinfo_documents = [
 
 autodoc_member_order = 'bysource'
 
-from _ext import supress_nonlocal_image_warn
+def supress_nonlocal_image_warn():
+    import sphinx.environment
+    sphinx.environment.BuildEnvironment.warn_node = _supress_nonlocal_image_warn
+
+def _supress_nonlocal_image_warn(self, msg, node):
+    from docutils.utils import get_source_line
+
+    if not msg.startswith('nonlocal image URI found:'):
+        self._warnfunc(msg, '%s:%s' % get_source_line(node))
+
 supress_nonlocal_image_warn()
